Simplifies handler dependency injection
