--- docs/architecture/params.md
@@ -1,3 +1,0 @@
-# auto_rest.params
-
-::: auto_rest.params
