+++ tests/cli/test_configure_cli_logging.py
