+++ tests/handlers/test_create_welcome_handler.py
