+++ tests/handlers/test_schema_handler.py
