+++ tests/models/test_create_session_iterator.py
