+++ tests/queries/test_delete_session_record.py
