+++ tests/queries/test_execute_session_query.py
