--- tests/unit_tests/params/test_create_pagination_dependency.py
@@ -1,81 +1,0 @@
-from unittest import TestCase
-
-from fastapi import FastAPI
-from fastapi.testclient import TestClient
-from sqlalchemy import Column, Integer, String
-from sqlalchemy.ext.declarative import declarative_base
-
-from auto_rest.params import create_pagination_dependency
-
-
-class DummyModel(declarative_base()):
-    """A dummy database model used for test fixtures."""
-
-    __tablename__ = "dummy_table"
-
-    id = Column(Integer, primary_key=True)
-    col1 = Column(String)
-
-
-class TestGetPaginationParams(TestCase):
-    """Unit tests for `create_pagination_dependency` function."""
-
-    @classmethod
-    def setUpClass(cls) -> None:
-        """Set up a FastAPI app and test client."""
-
-        app = FastAPI()
-
-        @app.get("/pagination")
-        def pagination(params: dict = create_pagination_dependency(DummyModel)):
-            return params
-
-        cls.client = TestClient(app)
-
-    def test_default_params(self) -> None:
-        """Verify default parameters are null."""
-
-        response = self.client.get("/pagination")
-        self.assertEqual({"limit": None, "offset": None}, response.json())
-
-    def test_negative_limit(self) -> None:
-        """Verify negative limit values fail validation."""
-
-        response = self.client.get("/pagination", params={"_limit_": -1})
-        error_detail = response.json()["detail"][0]
-
-        self.assertEqual(422, response.status_code)
-        self.assertEqual("Input should be greater than or equal to 0", error_detail["msg"])
-
-    def test_negative_offset(self) -> None:
-        """Verify negative offset values fail validation."""
-
-        response = self.client.get("/pagination", params={"_offset_": -1})
-        self.assertEqual(422, response.status_code)
-
-        error_detail = response.json()["detail"][0]
-        self.assertEqual("Input should be greater than or equal to 0", error_detail["msg"])
-
-    def test_zero_limit(self) -> None:
-        """Verify a zero limit value passes validation."""
-
-        response = self.client.get("/pagination", params={"_limit_": 0})
-        self.assertEqual({"limit": 0, "offset": None}, response.json())
-
-    def test_zero_offset(self) -> None:
-        """Verify a zero offset value passes validation."""
-
-        response = self.client.get("/pagination", params={"_offset_": 0})
-        self.assertEqual({"limit": None, "offset": 0}, response.json())
-
-    def test_positive_limit(self) -> None:
-        """Verify a positive limit value passes validation."""
-
-        response = self.client.get("/pagination", params={"_limit_": 200})
-        self.assertEqual({"limit": 200, "offset": None}, response.json())
-
-    def test_positive_offset(self) -> None:
-        """Verify a positive offset value passes validation."""
-
-        response = self.client.get("/pagination", params={"_offset_": 20})
-        self.assertEqual({"limit": None, "offset": 20}, response.json())
